+++ packages/pluggableWidgets/popup-menu-web/src/components/PopupTrigger.tsx
@@ -13,7 +13,12 @@ export const PopupTrigger = forwardRef(
                 className={"popupmenu-trigger"}
                 ref={ref}
                 data-state={open ? "open" : "closed"}
-                {...getReferenceProps?.()}
+                {...getReferenceProps?.({
+                    onClick: e => {
+                        e.preventDefault();
+                        e.stopPropagation();
+                    }
+                })}
             >
                 {children}
             </div>