+++ packages/pluggableWidgets/popup-menu-web/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   We fixed an issue where the user click interaction would trigger the overlayed element.
+
 ## [4.0.1] - 2024-12-13
 
 ### Breaking changes